// src/pages/BinanceLiveReal/components/PositionChartModal/types.ts
import { API } from '../../types';

export interface ChartProps {
  symbol: string;
  interval: string;
  entryPrice: number;
  entryTime: string | number;
  type: string;
  reason: string;
  strategy: string;
  gridUpper?: number;
  gridLower?: number;
  onClose: () => void;
}

export interface CandleData {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
}